requirements.txt
Pillow==9.5.0
albumentations
opencv-python
omegaconf
einops
transformers
timm
safetensors
easydict
xformers
pytorch-lightning
open_clip_torch
numpy
cython

fastapi
uvicorn
pydantic
requests
